Rename: api error 타입 파일명 수정
